// src/types/program-admin.ts
export interface ProgramAdmin {
  id: string;
  pgm_id: string;
  program_name: string;
  program_url: string;
  prifix: string;
  use_yn: string;
  program_type: string;
  search_yn: String;
  input_yn: string;
  delete_yn: string;
  excel_yn: string;
  print_yn: string;
  save_yn: string;
  user_name: string;
  isNew: boolean;
}